requirements.txt
healpy
matplotlib
numpy
scipy
numpy-stl
docopt
pytest